HSEARCH-3398 Move mapping/manager/search target of the ORM mapper to SPI
Users should not need these types to use the ORM mapper; the actual APIs
are the FullTextSession et. al.
