+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/session/spi/HibernateOrmSearchManager.java
@@ -4,11 +4,11 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.session;
+package org.hibernate.search.mapper.orm.session.spi;
 
 import java.util.Collection;
 
-import org.hibernate.search.mapper.orm.search.HibernateOrmSearchTarget;
+import org.hibernate.search.mapper.orm.search.spi.HibernateOrmSearchTarget;
 import org.hibernate.search.mapper.pojo.mapping.PojoWorkPlan;
 
 public interface HibernateOrmSearchManager extends AutoCloseable {
